Change VelocidadMaquina type to decimal(10,2) and add integer validation for PrdIlKwang

+++ Application/DTOs/PrdIlKwangDTO.cs
@@ -32,7 +32,7 @@ namespace Application.DTOs
 
         public int? NumeroPedido { get; set; }
 
-        public int VelocidadMaquina { get; set; }
+        public decimal VelocidadMaquina { get; set; }
 
         public int IdUbicacionBobinaA { get; set; }
